Add support for heterogeneous clusters
The Cluster section gains a new key, `partitions`, that gives a list of
partition names. Config sections with these names will be interpreted as
sub-clusters that inherit from the main Cluster section. This can be
useful in situations where a single cluster consists of multiple
generations of machines, each with a different amount of memory, number
of CPUs, number of GPUs, etc.