Evict jobs at end of glidein's allotted time (#166)
condor_master -r TIME causes condor_master to gracefully shut down after TIME, but TIME is in minutes, not seconds. With the correct units, and MaxJobRetirementTime set to the glidein's time to live, jobs are evicted before the parent batch system gets to them.

Also, dump logs to stdout if log shipping is not enabled, for easier local debugging.